fix: default values for all VmConfig except public_ip, move linux-specific API to a separate mod
